% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_color_map}
\alias{add_color_map}
\title{Add a new color map.}
\arguments{
\item{name}{a character string specifying the name of the new color map.}

\item{content}{a named vector of colors stored as character strings.}
}
\description{
Add a new color map.
}
