% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{Antler-class}
\alias{Antler-class}
\alias{Antler}
\title{A Reference Class to interface the Antler package.}
\description{
TODO
}
\section{Fields}{

\describe{
\item{\code{output_folder}}{a character string indicating the output directory.}
}}


