% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{brew_more_colors}
\alias{brew_more_colors}
\title{Build and name color palette vectors}
\usage{
brew_more_colors(ids, palette_name = rownames(RColorBrewer::brewer.pal.info))
}
\arguments{
\item{ids}{character or integer vector. The names that will be associated with the colors.}

\item{palette_name}{character string. The name of the RColorBrewer palette that will be extrapolated if needed. See \code{RColorBrewer::brewer.pal.info} for list of available colors.}
}
\description{
Build and name color palette vectors
}
