% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_pseudotime_tree}
\alias{build_pseudotime_tree}
\title{Reconstruct pseudotime tree}
\arguments{
\item{ordering}{character string. The name of the phenotypic metadata use to define the pseudotime ordering. Default to 'timepoint'.}

\item{process_plots}{logical value. Whether to show processing plots. Default to TRUE.}

\item{basename}{a character string indicating the path and prefix of the processing plots. Default to 'Pt_tree'.}
}
\description{
This function create a pseudotime tree structure and position each cell on one of its branches.
}
