% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{carve}
\alias{carve}
\title{Reconstruct cell state lineage trajectories}
\arguments{
\item{gene_modules_name}{character string. The name of the list of gene modules whose weights will be optimized by this method.}

\item{optim_strategy}{character string. Specifies the algorithm used to optimize the gene module weights. Either "SA" for simulated annealing (default) or "GA" for genetic algorithm. Different processing plots will be generated depending of the selected strategy.}

\item{optim_objectives}{a character vector containing a subset of of either:
\describe{
   \item{'timecorr'}{to favor a population of candidate cell lineages maximizing their Spearman correlation with the 'timepoint' phenotypic metadata.}
   \item{'complexity'}{to favor a population of candidate cell lineages minimizing the number of leaves of the reconstructed tree.}
   \item{'smoothness'}{to favor a population of candidate cell lineages minimizing the irregularity of the gene expression pattern along the differentiation trajectories.}
}
Multiple elements can be specified if the optimization strategy is set to 'GA' but a single one 'SA'. Default to \code{c('timecorr')}.}

\item{GA_selection_objectives}{character vector. If 'GA' is selected as optimization strategy, the final objectives used to select the optimal cell lineage. Any of the \code{optim_objectives} options can be selected. If more than one objective is specified, the geometric average of these objectives will be used to select the optimal solution. Default to "timecorr".}

\item{GA_num_generations}{integer value. If 'GA' is selected as optimization strategy, the number of generation of the genetic algorithm. Default to 200.}

\item{GA_pop_size}{integer value. If 'GA' is selected as optimization strategy, the size of population of the genetic algorithm. Default to 200.}

\item{SA_control}{List. If 'SA' is selected as optimization strategy, the list controlling the simulated annealing algorithm behavior. See documentation of the underlying \code{\link[GenSA]{GeneSA}} function.}

\item{SA_maxit}{interger value, the maximum number of iteration of the simulated annealing algorithm. Default to 500.}

\item{cluster_size_limit}{integer value. The maximum cell cluster size. If NULL (default) the \code{cluster_size_limit_ratio} argument is used instead.}

\item{cluster_size_limit_ratio}{numeric value. Specifies the maximum cell cluster size as a ratio of the total number of cells. Default to 0.02.}

\item{clustering_timepoint_first}{logical. Whether to pre-cluster the cells by 'timepoint' or not (default).}

\item{canalize}{logical. Whether to constraint the cell state graph using a diffusion kernel obtained from the optimal cluster level graph. Default to TRUE.}

\item{kernel_sigma}{numeric value. Defines the standard deviation of the diffusion kernel used to canalize the trajectories. Default to 1.}

\item{allow_multiple_roots}{(Experimental) logical. Whether to allow reconstructed cell lineages to start from multiple root states. Default to FALSE.}

\item{output_name}{character string. The name of the directory storing the processing plots. This directory will be created in the Antler's object output directory.}

\item{cell_colors}{character string. The name of the phenotypic metadata to show on the cell state graph plots. Default to 'cells_samples'.}

\item{seed}{integer value. Set the random number generator seed for reproducibility.}
}
\description{
See carving cell trajectories vignette.
}
