% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_clusters.R
\name{cell_clusters_identify}
\alias{cell_clusters_identify}
\title{Cluster cells from a selected gene module list.}
\arguments{
\item{name}{character string. The name of the cell clusters object. Multiple cell clusters objects under different names can be stored.}

\item{gene_modules_name}{character string. The name of the list of gene modules from which the cell-to-cell distance will be calculated.}

\item{method}{character string. The clustering method to use. Currently only \code{\link[stats]{hclust}} is available (using "ward.D2" as agglomeration method).}

\item{cell_cell_dist}{numeric matrix. Optionally a matrix containing the cell-to-cell distances can be provided. If NA (default), the matrix is calculated internally using Euclidean distance.}

\item{logscaled}{logical. If TRUE (default), the gene expression level are log-transformed and z-normalized before calculating the cell-to-cell distance. If FALSE, no transformation is applied to the levels.}

\item{num_clusters}{an integer value indicating the number of cell clusters. If NULL (default) a heuristic method is used to find the optimal number of clusters.}

\item{data_status}{character string. Specifies whether the gene expression levels to be used are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Normalized".}

\item{process_plots}{a logical value indicating whether to render the heuristic trade-off determining the number clusters. Default to FALSE.}
}
\description{
A text file containing the gene modules is written to the Antler output folder.
}
