% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_state_graph.R
\name{cell_state_graph_plot}
\alias{cell_state_graph_plot}
\title{Plot cell state graph}
\arguments{
\item{cell_colors}{character vector indicating the metadata to use to color the vertices. Names can be either one of the column names of the phenotypic metadata structure (see \code{colnames(pData(antler$expressionSet))}), names of a cluster entries, or gene names. The element of the character vector can be named themselves, in that case the element name will serve as a title for phenotypic metadata or cluster entries. If gene names are specified, their vector names must indicate the type of read counts that will be used to render the expression levels: "Raw", "Normalized" or "Smoothed". If no gene name's name is specified raw level will be rendered.}

\item{labels}{character vector. Labels to draw on top of the cells. Default to NA.}

\item{vertex.size}{numeric vector indicating the size of the vertices representing the cells.}

\item{vertex.shape}{character vector indicating the shape of the vertices. Can be either "circle" (default), "square" or "triangle".}

\item{vertex.frame.color}{character vector indicating the color of the frame of each vertex. Default to NA.}

\item{edge.width}{numeric value indicating the width of graph edges. Default to a fifth of the vertex size.}

\item{edge.color}{character string indicating the color of the edges. Default to "grey".}

\item{width}{numeric value indicating the width (in inches) of each individual subplots. Default to 3.}

\item{height}{numeric value indicating the height (in inches) of each individual subplots. Default to 3.}

\item{title}{logical value indicating whether to show each subplots' title. Default to TRUE.}

\item{save_pdf}{logical value indicating whether to save plot as a pdf file. Default to TRUE.}

\item{shuffle}{logical value indicating whether to randomize the z-order of each vertex. Default to FALSE.}

\item{vertex.plot.order}{integer vector. The z-order used to render the vertices. Default to NA.}

\item{gene_palette}{character string. The color palette to use if genes are specified as cell colors.}

\item{filename}{character string. The filename of the output file, if any.}

\item{...}{Extra arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Plot cell state graph
}
