% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert_ensembl_id_to_name}
\alias{convert_ensembl_id_to_name}
\title{Replace ensembl gene ids with gene names}
\arguments{
\item{biomart_dataset}{a character string specifying the BioMart dataset to use for the conversion. The list of available datasets can be listed with \link[biomaRt]{listDatasets} from the "biomaRt" package. Default to 'mmusculus_gene_ensembl'.}

\item{gene_ids_mapping_file}{a character string. If NULL, the conversion is performed from a gene ids/names mapping obtained using the biomaRt package (Default). This requires an internet connection. If not NULL, this argument must specify the location of a csv file with header containing the ensembl gene ids in the first column and the gene names in the second column.}
}
\description{
When the dataset is loaded with genes specified by their ensembl ids, this function maps these ids to their associated gene names.
}
