% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dispersed_genes}
\alias{dispersed_genes}
\title{Print highly-dispersed genes.}
\arguments{
\item{zscore_threshold}{a numeric value indicating the zcored dispersion threshold above which the gene names are returned. Default to 0.}

\item{num_bins}{a integer value indicating the number of bins used to calculate z-score into. Default to 20.}

\item{data_status}{character string. Specifies whether the gene expression levels used for calculation are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}

\item{invert}{logical. If FALSE (default) genes with z-scored dispersion higher than the threshold are returned. If FALSE, the complementary set is returned.}
}
\value{
a character vector of genes showing a sufficient level of dispersion.
}
\description{
The genes are allocated to a bin based on their average log-level of expression, then with each bin the dispersion (variance over mean of the log-levels) is z-scored. Genes having a sufficiently high dispersion z-scores are returned.
This method replicates \code{\link[Seurat]{FindVariableFeatures}} from the Seurat package.
}
