% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exclude_exotic_genes}
\alias{exclude_exotic_genes}
\title{Exclude genes from their name pattern.}
\arguments{
\item{pattern}{a character vector listing the patterns of gene names to exclude. Default to 'Gm[0-9].*', '.*-ps.*', 'ENSMUSG.*', '.*Rik' and '.*-rs*'.}

\item{verbose}{a logical indicating whether to print the list of excluded genes. Default to TRUE.}
}
\description{
Exclude genes from their name pattern.
}
