% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exclude_unexpressed_genes}
\alias{exclude_unexpressed_genes}
\title{Exclude genes with low expression.}
\arguments{
\item{min_cells}{a integer indicating the minimal number of cells a gene must be expressed in. Default to 3.}

\item{min_level}{a numeric value indicating the minimal level of expression required. Default to 0.}

\item{verbose}{a logical indicating whether to print the list of excluded genes. Default to FALSE.}

\item{data_status}{character string. Specifies whether the gene expression levels used for calculation are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}
}
\description{
Genes must be expressed in a given number of cell with a certain level to be kept.
}
