% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{export_expressionset}
\alias{export_expressionset}
\title{Export internal expression set.}
\arguments{
\item{exported_dataset.}{character string. The name of the directory containing the exported dataset. Default to 'exported_dataset'.}
}
\description{
The exported dataset is written in the Antler object's output directory.
}
