% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_modules.R
\name{gene_modules_export}
\alias{gene_modules_export}
\title{Save gene module to file.}
\arguments{
\item{name}{character string. The name of the list of gene modules that will be exported.}
}
\description{
A text file containing the gene modules is written to the Antler output folder.
}
