% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_modules.R
\name{gene_modules_report_genes}
\alias{gene_modules_report_genes}
\title{Report whether genes belong to modules}
\arguments{
\item{name}{character string. The name of the list of gene modules.}

\item{gene_names}{a character vector. Gene names to report.}
}
\description{
Report whether genes belong to modules
}
