% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_modules.R
\name{gene_modules_select_from_gene_names}
\alias{gene_modules_select_from_gene_names}
\title{Select gene modules from gene names}
\arguments{
\item{name_to}{character string. The name of the new list of gene modules.}

\item{name_from}{character string. The name of the original list from which gene modules are selected.}

\item{gene_names}{a character vector. Gene modules containing at least one of these gene names are kept.}
}
\description{
Select gene modules from gene names
}
