% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_modules.R
\name{gene_modules_select_from_ids}
\alias{gene_modules_select_from_ids}
\title{Select gene modules from indices}
\arguments{
\item{name_to}{character string. The name of the new list of gene modules.}

\item{name_from}{character string. The name of the original list from which gene modules are selected.}

\item{ids}{an integer vector. The indices of the gene modules to select.}
}
\description{
Select gene modules from indices
}
