% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{genes_from_GOterms}
\alias{genes_from_GOterms}
\title{Convert GO terms to gene names.}
\usage{
genes_from_GOterms(
  go_ids,
  biomart_dataset = "mmusculus_gene_ensembl",
  gofilename = system.file("extdata", "Annotations/gene_goterms.txt", package =
    "Antler")
)
}
\arguments{
\item{go_ids}{a character vector containing the GO terms to convert. Each term must be written with the 'GO:' prefix, e.g. 'GO:0007399'.}

\item{biomart_dataset}{a character string specifying the BioMart dataset to use for the conversion. The list of available datasets can be listed with \code{\link[biomaRt]{listDatasets}} from the "biomaRt" package. Default to 'mmusculus_gene_ensembl'.}

\item{gofilename}{character string. The file path of a csv file contining a mapping between gene names (first column) and GO terms (second column). If NA, the mapping is downloaded from ensembl's bioMart.}
}
\description{
This function download or load Genes/GO terms association maps from BioMart.
}
