% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_dataset}
\alias{load_dataset}
\title{Import a RNA-seq dataset into Antler.}
\arguments{
\item{folder_path}{character string. The path of the directory in which the dataset files are stored. if NULL (default), the "assayData" and "phenoData" arguments must be specified.}

\item{assayData_filename}{a character string specifying the name of the file containing the expression values.}

\item{phenoData_filename}{character string. The name of the file containing the cell metadata.}

\item{featData_filename}{a character string specifying the name of the file containing the gene metadata.}

\item{assayData}{Matrix. A matrix storing the expression values.}

\item{phenoData}{Matrix. A matrix storing the sample metadata.}

\item{data_status}{character string. Specifies whether the loaded expression are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}

\item{default_gene_names}{character string. Label describing the source of the gene names. Default to "ensembl_gene_id".}
}
\description{
Import a RNA-seq dataset into Antler.
}
