% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_pseudotime_dynamics}
\alias{plot_pseudotime_dynamics}
\title{Visualize the smoothed gene expression dynamics along specific branches of the pseudotime tree.}
\arguments{
\item{gene_list}{character vector. The names of the genes to render.}

\item{leaf_ids}{integer vector. Specifies the indices of the differentiation trajectories to visualize. If NULL (default), all the trajectories are rendered.}

\item{title}{character string. The title of the plot. Default to the rendered gene names.}

\item{subtitle}{character string. The subtitle of the plot. Default to NULL.}

\item{data_status}{character string. Specifies whether the gene expression levels to be rendered are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Smoothed".}

\item{show_cells}{logical. Whether to show the gene expression levels for each individual cells. Default to FALSE.}

\item{basename}{character string. The name of the output file, without extension.}
}
\description{
Visualize the smoothed gene expression dynamics along specific branches of the pseudotime tree.
}
