% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_QC}
\alias{plot_QC}
\title{Render various plots indicating the tallied gene and cell count distributions.}
\arguments{
\item{data_status}{a character string specifying whether the considered expression levels should be raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}

\item{basename}{a character string specifying the prefix of the rendered plots. Defaults to NULL.}

\item{reads_type}{a character string indicating whether the internal dataset stores generic "read" (default) or "UMI" elements.}

\item{feature_1}{an optional character string indicating a metadata feature used to aggregate the cells.}

\item{feature_2}{an optional character string indicating a metadata feature used to aggregate the cells.}
}
\value{
write plots as pdf files in \strong{Antler}'s \code{output_folder} output directory.
}
\description{
Render various plots indicating the tallied gene and cell count distributions.
}
