% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_counts}
\alias{read_counts}
\title{Returns a gene expression matrix.}
\arguments{
\item{data_status}{a character string specifying whether the returned expression levels are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}
}
\value{
a Matrix containing the gene expression levels with genes as rows and cells as columns.
}
\description{
Returns a gene expression matrix.
}
