% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_cells}
\alias{remove_cells}
\title{Exclude cells from the current dataset}
\arguments{
\item{ids}{an integer vector storing the indices of the  cells to remove.}
}
\description{
Exclude cells from the current dataset
}
