% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_clusters}
\alias{remove_clusters}
\title{Exclude cells belonging to one or more clusters}
\arguments{
\item{cell_clusters_name}{character string. The name of the cell clusters.}

\item{ids}{an integer vector storing the indices of the clusters to remove.}
}
\description{
If some genes were expressed only in the excluded cells, they will also be removed from the dataset.
}
