% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_genes}
\alias{remove_genes}
\title{Exclude genes from the current dataset.}
\arguments{
\item{ids}{an integer vector specifying the indices of the genes to remove.}

\item{names}{a character vector specifying the names of the genes to remove.}
}
\description{
Exclude genes from the current dataset.
}
