% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outlier genes and cells.}
\arguments{
\item{lowread_thres}{numeric value. Cells will be excluded if their total expression level is below this threshold. Default to -Inf.}

\item{lowread_high}{numeric value. Cells will be excluded if their total expression level is higher than this threshold. Default to Inf.}

\item{min_cells}{integer value. Genes expressed in less than this value will be excluded.}

\item{min_genes}{integer value. If a cell has a total number of expressed genes lower than this value, it is excluded.}

\item{verbose}{a logical indicating whether to print the dataset dimension at each filtering step. Default to TRUE.}

\item{data_status}{character string. Specifies whether the gene expression levels used for calculation are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}
}
\description{
This method excludes cells with a total read counts outside of a specified interval, cells having less than a certain number of expressed genes, and genes expressed in less than a given number of cells.
}
