% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_size}
\alias{sample_size}
\title{Returns a contingency table with the counts per type of cell samples.}
\value{
a table.
}
\description{
Returns a contingency table with the counts per type of cell samples.
}
