% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save_dataset}
\alias{save_dataset}
\title{Save dataset to file.}
\arguments{
\item{folder_path}{character string. The path of the directory in which the csv files are written.}

\item{data_status}{character string. Specifies whether the exported expression levels are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Raw".}
}
\description{
Export the gene expression matrix and the table of phenotypic sample information as csv files.
}
