% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_favorite_genes}
\alias{set_favorite_genes}
\title{Store a list of favorite gene names.}
\arguments{
\item{gene_list}{a character containing gene names.}
}
\description{
Store a list of favorite gene names.
}
