% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smooth}
\alias{smooth}
\title{Smooth gene expression using the cell state graph}
\arguments{
\item{method}{a character string among:
\describe{
   \item{'average'}{to smooth by averaging gene expression over the cell neighbors of the cell state graph (default).}
   \item{'diffusion'}{to smooth by diffusing the gene expression levels on the cell state graph.}
}}

\item{from}{character string. Specifies whether the original gene expression levels used are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Normalized".}
}
\description{
This function proposes two methods to smooth the gene expression levels: a simple average over the cells' local neighborhood or a more advanced diffusion-based imputation inspired by the "Rmagic" package.
}
